split axios functions in favoritePage into separated file (functions) and make style responsive relates #52
