+++ client/src/container/FavoritePage/AddOrRemoveFolder.js
@@ -1,7 +1,17 @@
 import React from 'react';
 import { Row, Col, Button } from 'antd';
 
-const AddOrRemoveFolder = ({ data, folderId, type, addOrRemoveFun }) => {
+import { addCourseToFolder, removeCourseFromFolder } from './functions';
+
+const AddOrRemoveFolder = ({
+  data,
+  folderId,
+  type,
+  allFavoriteData,
+  setDisplayFavoriteData,
+  displayFavoriteData,
+  setAllFavoriteData,
+}) => {
   const condition = (course) =>
     type === 'Add'
       ? course.folder_id !== Number(folderId)
@@ -19,12 +29,39 @@ const AddOrRemoveFolder = ({ data, folderId, type, addOrRemoveFun }) => {
             <Row style={{ width: '100%' }}>
               <Col span={18}>{course.title.substring(0, 50)}</Col>
               <Col span={6}>
-                <Button
-                  type="primary"
-                  onClick={() => addOrRemoveFun(course.course_id, folderId)}
-                >
-                  {type}
-                </Button>
+                {type === 'Add' ? (
+                  <Button
+                    type="primary"
+                    onClick={() =>
+                      addCourseToFolder(
+                        course.course_id,
+                        folderId,
+                        allFavoriteData,
+                        setDisplayFavoriteData,
+                        displayFavoriteData,
+                        setAllFavoriteData
+                      )
+                    }
+                  >
+                    {type}
+                  </Button>
+                ) : (
+                  <Button
+                    type="primary"
+                    onClick={() =>
+                      removeCourseFromFolder(
+                        course.course_id,
+                        folderId,
+                        allFavoriteData,
+                        setDisplayFavoriteData,
+                        displayFavoriteData,
+                        setAllFavoriteData
+                      )
+                    }
+                  >
+                    {type}
+                  </Button>
+                )}
               </Col>
             </Row>
           ))}
